+++ disperser/cmd/apiserver/lib/apiserver.go
@@ -157,7 +157,7 @@ func RunDisperserServer(ctx *cli.Context) error {
 			meterer,
 			authv2.NewPaymentStateAuthenticator(config.AuthPmtStateRequestMaxPastAge, config.AuthPmtStateRequestMaxFutureAge),
 			prover,
-			uint64(config.MaxNumSymbolsPerBlob),
+			uint32(config.MaxNumSymbolsPerBlob),
 			config.OnchainStateRefreshInterval,
 			logger,
 			reg,
